Revert most of the series for adding vexpress_aemv8r support
Per a request from Andre Przywara and agreed with by Peter Hoyes, the
vexpress aemv8r support wasn't quite ready to be merged, but the
discussion had moved off list.  We should keep the first patch in the
series for now, but revert the rest.  This reverts the following
commits:

10c283cdc80f doc: Add documentation for the Arm vexpress board configs
164ac9e2c20c arm: Use armv8_switch_to_el1 env to switch to EL1
94049432d0b4 vexpress64: Add BASER_FVP vexpress board variant
903989cccc7e armv8: Add ARMv8 MPU configuration logic
26ba79744f0e armv8: Ensure EL1&0 VMSA is enabled

Signed-off-by: Tom Rini <trini@konsulko.com>